requirements.txt
pydantic>=2.6,<3
PyYAML>=6.0.1
requests>=2.31
jsonschema>=4.21
